Fix dollarQuotedString() to handle dollar quotes in text
Fixes a bug in dollarQuotedString() where it fails to account for the
case where the string being quoted contains the randomly chosen dollar
quote tag. Before the fix, if a string containing $a$ there is a one in
21 (number of values in randomTags) chance of it occurring.

With the fix the function now checks if the chosen dollar quote tag is
contained within the string to be quoted and if so chooses a new longer
random tag. The tag initially starts as the empty string so the initial
default choice for quoting becomes $$.

Random additions to the tag were chosen, over incrementing a counter, to
make it slightly more difficult for an attacker to supply an input string
that would require a large number of iterations.

As the new defualt is to use $$ for quoting, a number of the results for
unit tests have been simplified to match on the exact text that would be
returned.