Expand set of random tag chars for dollar quotes
Change the set of characters used to generate random tags for dollar
quotes to include both lower and upper case non-vowel characters and
the digits two (2) through nine (9).

The overall expanded set decreases the likelyhood of a randomly picked
tag existing in the contained input.

The exclusions are to ensure that generated tags do not randomly end
up including naughty words.